quantum-voting/backend/requirements.txt
fastapi
uvicorn
qiskit
qiskit-aer
pycryptodome
pydantic
websockets
